Update service.ts
If the user calls the function sendPing the promise is never resolved as the call back in case of success is never called. 
Currently the callback is only called on case of failure (function rpcFailure) leading to a promise rejection. Adding the code in handlePong should resolve this.

F.e. if the user calls sendPing:
const pong = await this.discv5.sendPing(enr);